コンポーネントのディレクトリ構造変更
