+++ src/components/organisms/controls/Input.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
-import { computed, watch, ref } from 'vue';
+import { computed, watch } from 'vue';
 import { useField } from 'vee-validate';
 import { ZodString } from 'zod';
-import InputTextCounter from '@/components/organisms/formParts/InputTextCounter.vue';
+import InputTextCounter from '@/components/organisms/controls/InputTextCounter.vue';
 
 const model = defineModel<string>();
 const props = withDefaults(
@@ -28,6 +28,10 @@ const props = withDefaults(
          */
         disabled?: boolean;
         /**
+         * 種類
+         */
+        type?: 'text' | 'email' | 'password';
+        /**
          * 表示種類
          */
         valiant?: 'primary' | 'secondary' | 'info' | 'success' | 'warning' | 'danger';
@@ -35,18 +39,6 @@ const props = withDefaults(
          * サイズ
          */
         size?: 'small' | 'medium' | 'large';
-        /**
-         * デフォルト行数
-         */
-        line?: number;
-        /**
-         * 最小行数
-         */
-        minLine?: number | null;
-        /**
-         * 最大行数
-         */
-        maxLine?: number | null;
         /**
          * エラーメッセージを表示するか
          */
@@ -58,9 +50,9 @@ const props = withDefaults(
         label: ' ',
         placeholder: '',
         disabled: false,
+        type: 'text',
         valiant: 'secondary',
         size: 'medium',
-        line: 3,
         isErrorMessage: true
     }
 );
@@ -81,10 +73,6 @@ const max = computed(
         )?.value || null
 );
 
-const cssMinLine = computed(() => `${props.minLine ?? props.line}lh`);
-const cssMaxLine = computed(() => (props.maxLine ? `${props.maxLine}lh` : null));
-const textareaRef = ref();
-
 watch(value, (v) => {
     model.value = v;
 });
@@ -93,22 +81,17 @@ watch(value, (v) => {
 if (value.value == null && model.value != null) {
     value.value = model.value;
 }
-
-watch(value, (value) => {
-    let lines = (value + '\n').match(/\n/g)?.length ?? 1;
-    textareaRef.value.style.height = lines + 'lh';
-});
 </script>
 
 <template>
-    <label class="component-textarea">
+    <label class="component-input">
         <InputTextCounter v-if="max" class="counter" :text="value" :max="max" />
-        <textarea
-            ref="textareaRef"
-            v-model="value"
-            class="textarea"
+        <input
+            v-model.trim="value"
+            class="input"
             :class="[valiant, size]"
             placeholder=" "
+            :type="type"
             :name="name"
             :required="isRequired"
             :disabled="disabled"
@@ -124,7 +107,7 @@ watch(value, (value) => {
 </template>
 
 <style scoped>
-.component-textarea {
+.component-input {
     position: relative;
     text-align: left;
     display: block;
@@ -136,7 +119,7 @@ watch(value, (value) => {
         right: 0;
     }
 
-    .textarea {
+    .input {
         display: flex;
         gap: 16px;
         align-items: center;
@@ -147,10 +130,6 @@ watch(value, (value) => {
         padding: 0 8px;
         border: 1px solid var(--color-theme-border);
         border-radius: 4px;
-        /* field-sizing: content; 後に登場予定。まだ未実装 */
-        resize: none;
-        min-height: v-bind(cssMinLine);
-        max-height: v-bind(cssMaxLine);
         background-color: var(--color-theme-bg-primary);
         transition:
             color 0.2s,
@@ -195,8 +174,8 @@ watch(value, (value) => {
             content: '*';
         }
     }
-    .textarea:not(:placeholder-shown) + .label-placeholder,
-    .textarea:focus + .label-placeholder {
+    .input:not(:placeholder-shown) + .label-placeholder,
+    .input:focus + .label-placeholder {
         top: -0.5em;
         left: 0em;
         font-size: var(--font-size-small);
