+++ src/stories/organisms/common/Button.stories.ts
@@ -1,4 +1,4 @@
-import Button from '@/components/organisms/formParts/Button.vue';
+import Button from '@/components/organisms/common/Button.vue';
 import type { Meta, StoryObj } from '@storybook/vue3';
 
 const meta: Meta<typeof Button> = {
