+++ src/stories/organisms/controls/Number.stories.ts
@@ -1,27 +1,30 @@
-import Input from '@/components/organisms/formParts/Input.vue';
+import Number from '@/components/organisms/controls/Number.vue';
 import useFormData from '@/composables/useFormData';
 import type { Meta, StoryObj } from '@storybook/vue3';
-import { string, object } from 'zod';
+import { number, object } from 'zod';
 
 const TEST_SCHEMA = object({
-    test: string().max(50).min(1)
+    test: number().max(50).min(1)
 }).required();
 
-const meta: Meta<typeof Input> = {
-    component: Input,
+const meta: Meta<typeof Number> = {
+    component: Number,
     render: (args) => ({
-        components: { Input },
+        components: { Number },
         setup() {
-            useFormData(TEST_SCHEMA, { test: 'ヤマダ タロウ' });
+            useFormData(TEST_SCHEMA, { test: 10 });
             return { args };
         },
-        template: '<Input v-bind="args" />'
+        template: '<Number v-bind="args" />'
     }),
     tags: ['autodocs'],
     args: {
-        modelValue: ''
+        modelValue: 0
     },
     argTypes: {
+        modelValue: {
+            control: { type: 'number' }
+        },
         // TODO: 現状lintエラーの回避策なし
         // @ts-ignore
         'onUpdate:modelValue': { action: 'onUpdate:modelValue' }
@@ -29,28 +32,28 @@ const meta: Meta<typeof Input> = {
 };
 
 export default meta;
-type Story = StoryObj<typeof Input>;
+type Story = StoryObj<typeof Number>;
 
 export const Default: Story = {};
 
 export const Label: Story = {
     args: {
-        modelValue: '太郎',
-        label: '名前'
+        modelValue: 0,
+        label: '数字'
     }
 };
 
 export const Placeholder: Story = {
     args: {
         ...Label.args,
-        placeholder: '山田 太郎'
+        placeholder: '0'
     }
 };
 
 export const Disabled: Story = {
     args: {
         ...Label.args,
-        placeholder: '山田 太郎',
+        placeholder: '0',
         disabled: true
     }
 };
@@ -60,6 +63,6 @@ export const Schema: Story = {
         ...Label.args,
         name: 'test',
         schema: TEST_SCHEMA.shape.test,
-        placeholder: '山田 太郎'
+        placeholder: '0'
     }
 };
