+++ src/stories/organisms/controls/Radio.stories.ts
@@ -1,4 +1,4 @@
-import Radio from '@/components/organisms/formParts/Radio.vue';
+import Radio from '@/components/organisms/controls/Radio.vue';
 import useFormData from '@/composables/useFormData';
 import type { Meta, StoryObj } from '@storybook/vue3';
 import { literal, object } from 'zod';
